+++ pyquil/api/_qam.py
@@ -17,11 +17,13 @@ import warnings
 
 from abc import ABC, abstractmethod
 from collections import defaultdict
-from typing import Sequence, Union, Optional
+from typing import Dict, Sequence, Union, Optional
 
-from rpcq.messages import ParameterAref
+import numpy as np
+from rpcq.messages import BinaryExecutableResponse, ParameterAref, PyQuilExecutableResponse
 
 from pyquil.api._error_reporting import _record_call
+from pyquil.experiment._main import TomographyExperiment
 
 
 class QAMError(RuntimeError):
@@ -40,23 +42,26 @@ class QAM(ABC):
     """
 
     @_record_call
-    def __init__(self):
+    def __init__(self) -> None:
         self.reset()
 
     @_record_call
-    def load(self, executable):
+    def load(self, executable: Union[BinaryExecutableResponse, PyQuilExecutableResponse]) -> "QAM":
         """
         Initialize a QAM into a fresh state.
 
         :param executable: Load a compiled executable onto the QAM.
         """
+        self.status: str
         if self.status == "loaded":
             warnings.warn("Overwriting previously loaded executable.")
         assert self.status in ["connected", "done", "loaded"]
 
-        self._variables_shim = {}
-        self._executable = executable
-        self._memory_results = defaultdict(lambda: None)
+        self._variables_shim: Dict[ParameterAref, Union[int, float]] = {}
+        self._executable: Optional[
+            Union[BinaryExecutableResponse, PyQuilExecutableResponse]
+        ] = executable
+        self._memory_results: Optional[Dict[str, np.ndarray]] = defaultdict(lambda: None)
         self.status = "loaded"
         return self
 
@@ -65,9 +70,9 @@ class QAM(ABC):
         self,
         *,
         region_name: str,
+        value: Union[int, float, Sequence[int], Sequence[float]],
         offset: Optional[int] = None,
-        value: Optional[Union[int, float, Sequence[int], Sequence[float]]] = None,
-    ):
+    ) -> "QAM":
         """
         Writes a value or unwraps a list of values into a memory region on
         the QAM at a specified offset.
@@ -83,30 +88,18 @@ class QAM(ABC):
         elif isinstance(value, Sequence):
             warnings.warn("offset should be None when value is a Sequence")
 
-        if isinstance(value, Sequence):
-            assert all(
-                isinstance(v, type(value[0])) for v in value
-            ), "Element of 'value' must be of uniform type"
-            if (
-                region_name in self._variables_shim.keys()
-                and len(value) > len(self._variables_shim[region_name]) - offset
-            ):
-                raise ValueError("Value sequence exceeds memory region size")
-
-            for index, v in enumerate(value):
-                if not isinstance(v, type(value[0])):
-                    raise TypeError("Value sequence is not of uniform type")
-
-                aref = ParameterAref(name=region_name, index=offset + index)
-                self._variables_shim[aref] = v
-        else:
+        if isinstance(value, (int, float)):
             aref = ParameterAref(name=region_name, index=offset)
             self._variables_shim[aref] = value
+        else:
+            for index, v in enumerate(value):
+                aref = ParameterAref(name=region_name, index=offset + index)
+                self._variables_shim[aref] = v
 
         return self
 
     @abstractmethod
-    def run(self):
+    def run(self) -> "QAM":
         """
         Reset the program counter on a QAM and run its loaded Quil program.
         """
@@ -115,7 +108,7 @@ class QAM(ABC):
         return self
 
     @_record_call
-    def wait(self):
+    def wait(self) -> "QAM":
         """
         Blocks until the QPU enters the halted state.
         """
@@ -124,7 +117,7 @@ class QAM(ABC):
         return self
 
     @_record_call
-    def read_memory(self, *, region_name: str):
+    def read_memory(self, *, region_name: str) -> np.ndarray:
         """
         Reads from a memory region named region_name on the QAM.
 
@@ -135,11 +128,11 @@ class QAM(ABC):
         :return: A list of values of the appropriate type.
         """
         assert self.status == "done"
-
+        assert self._memory_results is not None
         return self._memory_results[region_name]
 
     @_record_call
-    def read_from_memory_region(self, *, region_name: str):
+    def read_from_memory_region(self, *, region_name: str) -> np.ndarray:
         """
         Reads from a memory region named region_name on the QAM.
 
@@ -158,7 +151,7 @@ class QAM(ABC):
         return self.read_memory(region_name=region_name)
 
     @_record_call
-    def reset(self):
+    def reset(self) -> None:
         """
         Reset the Quantum Abstract Machine to its initial state, which is particularly useful
         when it has gotten into an unwanted state. This can happen, for example, if the QAM
@@ -167,6 +160,6 @@ class QAM(ABC):
         self._variables_shim = {}
         self._executable = None
         self._memory_results = defaultdict(lambda: None)
-        self._experiment = None
+        self._experiment: Optional[TomographyExperiment] = None
 
         self.status = "connected"